Modified strjoin && strlen to check for null

+++ ft_strlen.c
@@ -11,11 +11,10 @@
 /* ************************************************************************** */
 
 #include "libft.h"
-
+#include <stdio.h>
 size_t	ft_strlen(const char *s)
 {
-	int	i;
-
+	size_t i;
 	i = 0;
 	while (s && s[i])
 		i++;
